+++ src/test/java/com/epam/tc/lesson01/utils/AnotherTest.java
@@ -1,0 +1,10 @@
+package com.epam.tc.lesson01.utils;
+
+import org.testng.annotations.Test;
+
+public class AnotherTest {
+    @Test
+    public void emptyTest() {
+        //nop
+    }
+}
