--- src/test/java/com/epam/tc/utils/AnotherTest.java
@@ -1,10 +1,0 @@
-package com.epam.tc.utils;
-
-import org.testng.annotations.Test;
-
-public class AnotherTest {
-    @Test
-    public void emptyTest() {
-        //nop
-    }
-}
